+++ kernel/mm/palloc.c
@@ -418,7 +418,7 @@ void MmInitPhysicalAllocator(const void *bootArgs)
         {
             //reserve memory that stores kernel symbol and string tables
             const struct Multiboot2ElfSymbolsTag *elf = (const struct Multiboot2ElfSymbolsTag*)tag;
-            uint16_t count = elf->num; //get section header count
+            size_t count = elf->num; //get section header count
             const struct Elf32_Shdr *s = (const struct Elf32_Shdr*)(elf + 1);
             while(0 != count)
             {
